auth and todo routes

+++ tests/setup.js
@@ -1,16 +1,23 @@
 const mongoose = require("mongoose");
+require("dotenv").config();
 
 beforeAll(async () => {
-  await mongoose.connect(process.env.MONGO_URI);
-  console.log(process.env.MONGO_URI);
-  console.log("Connected To DB In Jest Config");
+  try {
+    await mongoose.connect(process.env.MONGO_URI);
+    console.log("Connected To DB In Jest Config");
+  } catch (error) {
+    console.log("Unable to connect DB", error);
+  }
 });
 
 afterAll(async () => {
-  const collections = await mongoose.connection.db.collections();
-  for (let collection of collections) {
-    await collection.deleteMany({});
+  if (mongoose.connection.readyState === 1) {
+    // connected
+    const collections = await mongoose.connection.db.collections();
+    for (let collection of collections) {
+      await collection.deleteMany({});
+    }
+    await mongoose.connection.close();
+    console.log("Collections cleared and connection closed");
   }
-  await mongoose.connection.close();
-  console.log("Collections cleared and connection closed");
 });
